refactor(compact_radar): split CompactRadarPanel into mixin package
Move widgets/panels/compact_radar_panel.py into a package: frame.py
builds the UI; properties.py holds the RadarViewMode enum, the top-cards
limits, and the view_mode getter; handlers.py owns the toggle callback
and the populate helpers.
