+++ widgets/panels/compact_radar_panel/handlers.py
@@ -1,85 +1,29 @@
-"""
-Compact Radar Panel - Displays archetype card frequency in a small format.
-
-Designed for embedding in the opponent tracker overlay.
-Two view modes:
-  - "Top Cards": Top 15 mainboard / 8 sideboard cards with inclusion rates
-  - "Full Decklist": All cards as an average decklist (avg copies rounded)
-"""
+"""Event handlers, public state setters, and list populators for the compact radar panel."""
 
 from __future__ import annotations
-
-from enum import Enum
 
 import wx
 from loguru import logger
 
 from services.radar_service import RadarData
-from utils.constants import DARK_BG, DARK_PANEL, LIGHT_TEXT, SUBDUED_TEXT
+from widgets.panels.compact_radar_panel.properties import (
+    _TOP_MAINBOARD_LIMIT,
+    _TOP_SIDEBOARD_LIMIT,
+    RadarViewMode,
+)
 
 
-class RadarViewMode(Enum):
-    """View modes for the compact radar panel."""
+class CompactRadarHandlersMixin:
+    """Public API, toggle callback, and list population for :class:`CompactRadarPanel`."""
 
-    TOP_CARDS = "top"
-    FULL_DECKLIST = "full"
-
-
-_TOP_MAINBOARD_LIMIT = 15
-_TOP_SIDEBOARD_LIMIT = 8
-
-
-class CompactRadarPanel(wx.Panel):
-    """Compact panel for displaying radar data in small overlays."""
-
-    def __init__(self, parent: wx.Window):
-        super().__init__(parent)
-        self.SetBackgroundColour(DARK_PANEL)
-
-        self.current_radar: RadarData | None = None
-        self._view_mode: RadarViewMode = RadarViewMode.TOP_CARDS
-
-        self._build_ui()
-        self.Hide()
-
-    def _build_ui(self) -> None:
-        sizer = wx.BoxSizer(wx.VERTICAL)
-        self.SetSizer(sizer)
-
-        # Header row: label + view toggle button
-        header_sizer = wx.BoxSizer(wx.HORIZONTAL)
-        sizer.Add(header_sizer, 0, wx.EXPAND | wx.ALL, 4)
-
-        self.header_label = wx.StaticText(self, label="Radar: Loading...")
-        self.header_label.SetForegroundColour(LIGHT_TEXT)
-        font = self.header_label.GetFont()
-        font = font.Bold()
-        self.header_label.SetFont(font)
-        header_sizer.Add(self.header_label, 1, wx.ALIGN_CENTER_VERTICAL)
-
-        self.view_toggle_btn = wx.Button(self, label="Full Decklist", size=(90, 22))
-        self.view_toggle_btn.SetBackgroundColour(DARK_BG)
-        self.view_toggle_btn.SetForegroundColour(LIGHT_TEXT)
-        self.view_toggle_btn.Bind(wx.EVT_BUTTON, self._on_toggle_view)
-        self.view_toggle_btn.Hide()
-        header_sizer.Add(self.view_toggle_btn, 0, wx.LEFT, 4)
-
-        # Status label (for loading/errors)
-        self.status_label = wx.StaticText(self, label="")
-        self.status_label.SetForegroundColour(SUBDUED_TEXT)
-        sizer.Add(self.status_label, 0, wx.LEFT | wx.RIGHT | wx.BOTTOM, 4)
-
-        # Scrollable list for cards
-        self.card_list = wx.ListBox(self, style=wx.LB_SINGLE)
-        self.card_list.SetBackgroundColour(DARK_BG)
-        self.card_list.SetForegroundColour(LIGHT_TEXT)
-        sizer.Add(self.card_list, 1, wx.EXPAND | wx.LEFT | wx.RIGHT | wx.BOTTOM, 4)
+    current_radar: RadarData | None
+    _view_mode: RadarViewMode
+    header_label: wx.StaticText
+    view_toggle_btn: wx.Button
+    status_label: wx.StaticText
+    card_list: wx.ListBox
 
     # ============= Public API =============
-
-    @property
-    def view_mode(self) -> RadarViewMode:
-        return self._view_mode
 
     def display_radar(self, radar: RadarData) -> None:
         self.current_radar = radar
